+++ app/IOT/python/config.ini
@@ -5,14 +5,18 @@ topic_subscribe = solaredge/#,AM107/by-room/
 
 [DATA]
 salles = 
-types = temperature,humidity,room,co2
-typesSolaire = lifeTimeData,lastUpdateTime
-frequency = 3600
-threshold_temperature = 25
-threshold_co2 = 2000
-threshold_humidity = 55
+types = temperature,humidity,co2,room
+typesSolaire = lifeTimeData,lastUpdateTime,currentPower
+frequency = 120
+
 
 [OUTPUT]
-json_file = donneeJSON.json
+json_file = dataJSON.json
 alert_file = alertJSON.json
 
+[ALERTE]
+seuil_temperature = 25
+seuil_co2 = 1500
+seuil_humidity = 60
+seuil_power = 3000
+seuil_energy = 10000
